refactor(email settings): extract language settings button
